telegram_bot/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "faster-whisper>=1.0.0",
]

[tool.setuptools]
py-modules = ["bot", "handlers", "api_client", "config"]

